Cleaning up test organization for impending move
I'm moving/deleting some tests to make an upcoming RESTClient refactor
PR cleaner.

 - TestDoRequestBearer is redundant to
   pkg/client/transport.TestBearerAuthRoundTripper
 - Added a case to pkg/client/transport.TestBasicAuthRoundTripper to
   cover what TestDoRequestWithoutPassword was testing
 - Moved TestSetsCodec, TestRESTClientRequires,
   TestValidateHostParameter to helper_test.go since they were really
   testing helper functions and not the RESTClient directly.
 - Modified TestValidateHostParameter during the moved to just use
   DefaultServerURL instead of creating a client.
